busybox: put execvp into a header file
And re-generate the .i files.
