+++ c/busybox-1.22.0/chroot-incomplete_true-no-overflow_true-valid-memsafety.c
@@ -89,25 +89,6 @@ static const char *msg_eol = "\n";
 // file libbb/default_error_retval.c line 18
 static unsigned char xfunc_error_retval = (unsigned char)1;
 
-int execvp(const char *file, char * const *argv)
-{
-	int i;
-	/* go through the arguments to "check" that they
-	   are null-terminated */
-	for (i = 0; argv[i] != 0; ++i)
-		;
-
-	if (__VERIFIER_nondet_int()) {
-		*bb_errno = __VERIFIER_nondet_int();
-		__VERIFIER_assume(*bb_errno != 0);
-		return -1;
-	}
-
-	/* exec succeeded, exit the original program
-	 * (use abort because of memory leaks) */
-	abort();
-}
-
 signed int find_applet_by_name(const char *applet) {
 	(void)applet;
 	return __VERIFIER_nondet_int();
@@ -393,5 +374,6 @@ static void xfunc_die(void)
   exit((signed int)xfunc_error_retval);
 }
 
+#include "busybox_sv_comp-execvp.h"
 #include "busybox_sv_comp-getpwnam.h"
 #include "busybox_sv_comp_impl.h"
